requirements.txt
discord.py==2.2.3
python-dotenv==1.0.0
Requests==2.29.0
requests_cache==1.0.1
yt_dlp==2023.3.4